==> TCC/1.PreparacaoDosDados.Rmd <==
---
output:
  word_document: default
  html_document: default
---

title: "Trabalho de Conclusão de Cusros - FGV T4 Big Data e Analytics"
author: "Ricardo Squassina Lee, Rafael Furlan, Djalma Gomes"
output:
  word_document: default
  html_notebook: default
  pdf_document: default
  html_document:
    df_print: paged



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r Inicia Ambiente, message=FALSE, include=FALSE}
#Iniciando o ambiente
rm(list = ls())
cat("\014") 
setwd("./")
```


```{r Carrega bibliotecas necessarias, message=FALSE, warning=TRUE, include=FALSE}
if(!require(data.table)) {
  install.packages("data.table")
  library(data.table)
}
if(!require(dummies)) {
  install.packages("dummies")
  library(dummies)
}
```


```{r Preparando os diretórios}
if(!dir.exists("./DADOS_ZIP")){dir.create("./DADOS_ZIP")}
if(!dir.exists("./DADOS")){dir.create("./DADOS")}
```



```{r CID-10}
# url <- "http://www.datasus.gov.br/cid10/V2008/downloads/CID10CSV.zip"
# dest <- "./DADOS_ZIP/CID10CSV.zip"
# download.file(url = url, destfile = dest)
# 
# unzip(dest, exdir = "./DADOS")
```


```{r TUSS}
# url <- "http://ans.gov.br/images/stories/Plano_de_saude_e_Operadoras/Area_do_consumidor/rol_tuss_rol_simplificado_rn305tuss_alterada_pela_rn_349_2014_ajustada_site_11-03-2015.xls"
# dest <- "./DADOS/tuss_22.xls"
# download.file(url = url, destfile = dest)
```



```{r Hospitalar SP, include=FALSE}
# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/HOSPITALAR/2015/SP.zip"
# dest <- "./DADOS_ZIP/SP-H-2015.zip"
# download.file(url = url, destfile = dest)
# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/HOSPITALAR/2016/SP.zip"
# dest <- "./DADOS_ZIP/SP-H-2016.zip"
# download.file(url = url, destfile = dest)
# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/HOSPITALAR/2017/SP.zip"
# dest <- "./DADOS_ZIP/SP-H-2017.zip"
# download.file(url = url, destfile = dest)

# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/AMBULATORIAL/2015/SP.zip"
# dest <- "./DADOS_ZIP/SP-A-2015.zip"
# download.file(url = url, destfile = dest)
# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/AMBULATORIAL/2016/SP.zip"
# dest <- "./DADOS_ZIP/SP-A-2016.zip"
# download.file(url = url, destfile = dest)
# url <- "http://ftp.dadosabertos.ans.gov.br/FTP/PDA/TISS/AMBULATORIAL/2017/SP.zip"
# dest <- "./DADOS_ZIP/SP-A-2017.zip"
# download.file(url = url, destfile = dest)

# do.call(rbind, lapply(list.files(path = "./DADOS_ZIP"
#                                  , pattern = "SP.*zip"
#                                  , full.names = TRUE
#                                  )
#                       , function(x) unzip(x, exdir = "./DADOS")))

```



```{r read csv files - Hospital Consolidado SP e CID10, echo=TRUE}
CONSULTAS <- do.call(rbind, lapply(list.files(path = "./DADOS"
                                    , pattern = ".*HOSP.*CONS.*csv"
                                    , full.names = TRUE)
                                    , function(x) fread(x,encoding = "Latin-1")))

# arquivos <- list.files(path = "./DADOS"
#                   , pattern = ".*CSV"
#                   , full.names = TRUE)

# CID10 <- lapply(arquivos, function(x) fread(x, encoding = "Latin-1", sep = ";"))
# names(CID10) <- substr(arquivos, 9, 32)
# rm(arquivos)

# CID10 <- fread("./DADOS/CID-10-CATEGORIAS.CSV", encoding = "Latin-1", sep = ";")

# TUSS_22 <- readxl::read_xls("./DADOS/tuss_22.xls", skip = 2)


```


```{r primeira visao}
summary(CONSULTAS)
nrow(CONSULTAS)
```

Código do Município de São Paulo: 355030
```{r Selecionando apenas o municipio de São Paulo}
CONSULTAS <- CONSULTAS[CONSULTAS$CD_MUNIC_BENEFICIARIO == 355030,]
```


```{r segunda visao}
summary(CONSULTAS)
nrow(CONSULTAS)
```

```{r terceira visao}
CONSULTAS <- within(CONSULTAS, rm("ID_PLANO"
                                    ,"CD_MUNIC_BENEFICIARIO"
                                    ,"PORTE_OPERADORA"
                                    ,"CD_MODALIDADE_OPERADORA"
                                    ,"MODALIDADE_OPERADORA"
                                    ,"CD_MUNIC_PRESTADOR"
                                    ,"UF_PRESTADOR"
                                    ,"DT_SAIDA_INTERNACAO"
                                    ,"CARATER_ATENDIMENTO"
                                    ,"TIPO_INTERNACAO"
                                    ,"REGIME_INTERNACAO"
                                    ,"MOTIVO_ENCERRAMENTO" 
                                    ,"NR_DIARIAS_ACOMPANHANTE"
                                    ,"NR_DIARIAS_UTI"
                                    ,"LG_VALOR_PREESTABELECIDO"))

summary(CONSULTAS)
nrow(CONSULTAS)
```

```{r NA treatment}
# CONSULTAS <- CONSULTAS[complete.cases(CONSULTAS),]
# CONSULTAS[!is.na(CONSULTAS$CID_1)]
# 
# summary(CONSULTAS)
# nrow(CONSULTAS)

```


```{r listing all the CIDs available in the datafram in a single column}
CID1 <- setnames(within(CONSULTAS[!which(CONSULTAS$CID_1 == "")], rm("CID_2","CID_3","CID_4")),"CID_1", "CID")
CID2 <- setnames(within(CONSULTAS[!which(CONSULTAS$CID_2 == "")], rm("CID_1","CID_3","CID_4")),"CID_2", "CID")
CID3 <- setnames(within(CONSULTAS[!which(CONSULTAS$CID_3 == "")], rm("CID_2","CID_1","CID_4")),"CID_3", "CID")
CID4 <- setnames(within(CONSULTAS[!which(CONSULTAS$CID_4 == "")], rm("CID_2","CID_3","CID_1")),"CID_4", "CID")

CONSULTAS <- data.table::rbindlist(l = list(CID1, CID2, CID3, CID4),use.names = TRUE)
rm(CID1, CID2, CID3, CID4)

CONSULTAS$CID <- substr(CONSULTAS$CID,1,3)

CONSULTAS$DT_INTERNACAO <- as.Date(CONSULTAS$DT_INTERNACAO, format = "%d/%m/%Y")
```

```{r visao final}
nrow(CONSULTAS)
length(unique(CONSULTAS$`#ID_EVENTO`))

```


```{r understanding the data  - HOSPITAL - SP - after NA treatment}
summary(CONSULTAS)
write.csv(CONSULTAS, file = "./consultas.csv")
```


```{r ajuste datas}
CONSULTAS <- fread("./consultas.csv", drop = "V1")
CONSULTAS$DT_INTERNACAO <- as.Date((CONSULTAS$DT_INTERNACAO))
CLIMA <- fread("./DADOS/base_saopaulo.csv", drop = c("V1","Estacao"))
CLIMA$Data <- as.Date(CLIMA$Data, format = "%m/%d/%Y")
```


```{r merge das tabelas}
CONSULTAS_CONS <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "Data")
head(CONSULTAS_CONS)
write.csv(CONSULTAS_CONS, file = "./consultas_cons.csv")
```

```{r}

# CONSULTAS_CONS <- fread("consultas_cons.csv", drop = c("V1"))
# 
# CONSULTAS_CONS2 <- cbind(CONSULTAS_CONS,dummy(CONSULTAS_CONS$CID))
# write.csv(CONSULTAS_CONS2, file = "consultas_cons_dummies.csv")

```

```{r organizando os capitulos}
CONSULTAS_CONS_CAP <- fread("consultas_cons.csv", drop = c("V1"))

CONSULTAS_CONS_CAP$CAP <- 99

CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "A", ][, "CAP"] <- 1
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "B", ][, "CAP"] <- 1
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "C", ][, "CAP"] <- 2
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "D" & substr(CONSULTAS_CONS_CAP$CID,2,2) < 5,][, "CAP"] <-2# < 5
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "D" & substr(CONSULTAS_CONS_CAP$CID,2,2) > 4,][, "CAP"] <-3 # > 4
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "E",][, "CAP"] <-4
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "F",][, "CAP"] <-5
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "G",][, "CAP"] <-6
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "H" & substr(CONSULTAS_CONS_CAP$CID,2,2) < 6,][, "CAP"] <-7# < 6
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "H" & substr(CONSULTAS_CONS_CAP$CID,2,2) > 5,][, "CAP"] <-8# > 5
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "I",][, "CAP"] <-9
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "J",][, "CAP"] <-10
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "K",][, "CAP"] <-11
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "L",][, "CAP"] <-12
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "M",][, "CAP"] <-13
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "N",][, "CAP"] <-14
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "O",][, "CAP"] <-15
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "P",][, "CAP"] <-16
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "Q",][, "CAP"] <-17
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "R",][, "CAP"] <-18
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "S",][, "CAP"] <-19
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "T",][, "CAP"] <-19
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "U",][, "CAP"] <-22
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "V",][, "CAP"] <-20
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "W",][, "CAP"] <-20
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "X",][, "CAP"] <-20
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "Y",][, "CAP"] <-20
CONSULTAS_CONS_CAP[substr(CONSULTAS_CONS_CAP$CID,1,1) == "Z",][, "CAP"] <-21

# tail(CONSULTAS_CONS_CAP)

CONSULTAS_CONS_CAP <- within(CONSULTAS_CONS_CAP, rm("V12"
                                    ,"V13"))
write.csv(CONSULTAS_CONS_CAP, file = "consultas_cons_cap.csv")

```


```{r}
CONSULTAS_CONS_CAP <- fread("consultas_cons_cap.csv", drop = c("V1"))

CONSULTAS_CONS_CAP2 <- cbind(CONSULTAS_CONS_CAP,dummy(CONSULTAS_CONS_CAP$CAP))
write.csv(CONSULTAS_CONS_CAP2, file = "consultas_cons_cap_dummies.csv")

```


```{r ajuste datas2 - data-x}
CONSULTAS <- fread("./consultas.csv", drop = "V1")
CONSULTAS$DT_INTERNACAO <- as.Date((CONSULTAS$DT_INTERNACAO))
CLIMA <- fread("./Rafael/DADOS/base_saopaulo.csv", drop = c("V1","Estacao"))
CLIMA$Data <- as.Date(CLIMA$Data, format = "%Y-%m-%d")
```


```{r merge das tabelas - data-7}
CLIMA$DATA_ANT <- CLIMA$Data + 7
CONSULTAS_CONS7 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS7 <- within(CONSULTAS_CONS7, rm("Data"))
write.csv(CONSULTAS_CONS7, file = "./consultas_cons7.csv")
```


```{r merge das tabelas - data-6}

CLIMA$DATA_ANT <- CLIMA$Data + 6
CONSULTAS_CONS6 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS6 <- within(CONSULTAS_CONS6, rm("Data"))
write.csv(CONSULTAS_CONS6, file = "./consultas_cons6.csv")
```


```{r merge das tabelas - data-5}

CLIMA$DATA_ANT <- CLIMA$Data + 5
CONSULTAS_CONS5 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS5 <- within(CONSULTAS_CONS5, rm("Data"))
write.csv(CONSULTAS_CONS5, file = "./consultas_cons5.csv")
```


```{r merge das tabelas - data-4}

CLIMA$DATA_ANT <- CLIMA$Data + 4
CONSULTAS_CONS4 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS4 <- within(CONSULTAS_CONS4, rm("Data"))
write.csv(CONSULTAS_CONS4, file = "./consultas_cons4.csv")
```


```{r merge das tabelas - data-3}

CLIMA$DATA_ANT <- CLIMA$Data + 3
CONSULTAS_CONS3 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS3 <- within(CONSULTAS_CONS3, rm("Data"))
write.csv(CONSULTAS_CONS3, file = "./consultas_cons3.csv")
```


```{r merge das tabelas - data-2}

CLIMA$DATA_ANT <- CLIMA$Data + 2
CONSULTAS_CONS2 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS2 <- within(CONSULTAS_CONS2, rm("Data"))
write.csv(CONSULTAS_CONS2, file = "./consultas_cons2.csv")
```


```{r merge das tabelas - data-1}

CLIMA$DATA_ANT <- CLIMA$Data + 1
CONSULTAS_CONS1 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS1 <- within(CONSULTAS_CONS1, rm("Data"))
write.csv(CONSULTAS_CONS1, file = "./consultas_cons1.csv")

```


```{r ajuste datas2 - CAP - data-x}
CONSULTAS <- fread("consultas_cons_cap.csv", drop = c("V1"))
CONSULTAS$DT_INTERNACAO <- as.Date((CONSULTAS$DT_INTERNACAO))
CLIMA <- fread("./DADOS/base_saopaulo.csv", drop = c("V1","Estacao"))
CLIMA$Data <- as.Date(CLIMA$Data, format = "%m/%d/%Y")
CONSULTAS <- CONSULTAS[,c(1:5,14)]
```


```{r merge das tabelas - CAP - data-7}
CLIMA$DATA_ANT <- CLIMA$Data + 7
CONSULTAS_CONS7 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS7 <- within(CONSULTAS_CONS7, rm("Data"))
write.csv(CONSULTAS_CONS7, file = "./consultas_cons_cap7.csv")
```


```{r merge das tabelas - CAP - data-6}

CLIMA$DATA_ANT <- CLIMA$Data + 6
CONSULTAS_CONS6 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS6 <- within(CONSULTAS_CONS6, rm("Data"))
write.csv(CONSULTAS_CONS6, file = "./consultas_cons_cap6.csv")
```


```{r merge das tabelas - CAP - data-5}

CLIMA$DATA_ANT <- CLIMA$Data + 5
CONSULTAS_CONS5 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS5 <- within(CONSULTAS_CONS5, rm("Data"))
write.csv(CONSULTAS_CONS5, file = "./consultas_cons_cap5.csv")
```


```{r merge das tabelas - CAP - data-4}

CLIMA$DATA_ANT <- CLIMA$Data + 4
CONSULTAS_CONS4 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS4 <- within(CONSULTAS_CONS4, rm("Data"))
write.csv(CONSULTAS_CONS4, file = "./consultas_cons_cap4.csv")
```


```{r merge das tabelas - CAP - data-3}

CLIMA$DATA_ANT <- CLIMA$Data + 3
CONSULTAS_CONS3 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS3 <- within(CONSULTAS_CONS3, rm("Data"))
write.csv(CONSULTAS_CONS3, file = "./consultas_cons_cap3.csv")
```


```{r merge das tabelas - CAP - data-2}

CLIMA$DATA_ANT <- CLIMA$Data + 2
CONSULTAS_CONS2 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS2 <- within(CONSULTAS_CONS2, rm("Data"))
write.csv(CONSULTAS_CONS2, file = "./consultas_cons_cap2.csv")
```


```{r merge das tabelas - CAP - data-1}

CLIMA$DATA_ANT <- CLIMA$Data + 1
CONSULTAS_CONS1 <- merge(CONSULTAS, CLIMA, by.x = "DT_INTERNACAO", by.y = "DATA_ANT")
CONSULTAS_CONS1 <- within(CONSULTAS_CONS1, rm("Data"))
write.csv(CONSULTAS_CONS1, file = "./consultas_cons_cap1.csv")

```
