==> TCC/2.AnaliseDosDados-MR-COR-VAR-COV-CAP.Rmd <==
---
title: "Trabalho de Conclusão de Cusros - FGV T4 Big Data e Analytics"
author: "Ricardo Squassina Lee, Rafael Furlan, Djalma Gomes"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
  word_document: default
---

---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


---
```{r Inicia Ambiente, message=FALSE, include=FALSE}
#Iniciando o ambiente
rm(list = ls())
cat("\014") 
# setwd("~/Tresors/zz-pessoal/FGV/git/TCC")
setwd("./")
```

---
```{r Carrega bibliotecas necessarias, message=FALSE, warning=TRUE, include=FALSE}
if(!require(data.table)) {
  install.packages("data.table")
  library(data.table)
}

```

```{r}
CONSULTAS_CONS <- fread("consultas_cons_cap.csv", drop = c("V1","V12","V13","CID1"))
```

# MapReduce

```{r MapReduce}
CAT_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Precip))
names(CAT_Precip_freq) <- c("CAP","Precip","Freq")
CAT_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$TMax, digits = 0)))
names(CAT_TMax_freq) <- c("CAP","TMax","Freq")
CAT_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$TMin, digits = 0)))
names(CAT_Tmin_freq) <- c("CAP","TMin","Freq")
CAT_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Ins))
names(CAT_Ins_freq) <- c("CAP","Ins","Freq")
CAT_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$EvapPiche))
names(CAT_EvapPiche_freq) <- c("CAP","EvapPiche","Freq")
CAT_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TempCompMedia))
names(CAT_TempCompMed_freq) <- c("CAP","TempCompMedia","Freq")
CAT_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$UmidRelMed))
names(CAT_UmidRelMed_freq) <- c("CAP","UmidRelMed","Freq")
CAT_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$VelVentMed))
names(CAT_VelVentMed_freq) <- c("CAP","VelVentMed","Freq")
CAT_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
names(CAT_TDelta_freq) <- c("CAP","TDelta","Freq")
CAT_TDeltaA_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round((CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin), digits = 0)))
names(CAT_TDeltaA_freq) <- c("CAP","TDelta","Freq")

# SEX_CAT_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Precip))
# names(SEX_CAT_Precip_freq) <- c("CAP","Sex","Freq")
# SEX_CAT_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,round(CONSULTAS_CONS$TMax, digits = 0)))
# names(SEX_CAT_TMax_freq) <- c("CAP","Sex","Tmax","Freq")
# SEX_CAT_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,round(CONSULTAS_CONS$TMin, digits =0 )))
# names(SEX_CAT_Tmin_freq) <- c("CAP","Sex","TMin","Freq")
# SEX_CAT_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Ins))
# names(SEX_CAT_Ins_freq) <- c("CAP","Sex","Ins","Freq")
# SEX_CAT_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
# names(SEX_CAT_EvapPiche_freq) <- c("CAP","Sex","EvapPiche","Freq")
# SEX_CAT_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
# names(SEX_CAT_TempCompMed_freq) <- c("CAP","Sex","TempCompMed","Freq")
# SEX_CAT_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
# names(SEX_CAT_UmidRelMed_freq) <- c("CAP","Sex","UmidRelMed","Freq")
# SEX_CAT_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
# names(SEX_CAT_VelVentMed_freq) <- c("CAP","Sex","VelVentMed","Freq")
# SEX_CAT_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
# names(SEX_CAT_TDelta_freq) <- c("CAP","Sex","TDelta","Freq")
# 
# IDADE_CAT_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Precip))
# names(IDADE_CAT_Precip_freq) <- c("CAP","Idade","Precip","Freq")
# IDADE_CAT_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,round(CONSULTAS_CONS$TMax, digits = 0)))
# names(IDADE_CAT_TMax_freq) <- c("CAP","Idade","TMax","Freq")
# IDADE_CAT_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,round(CONSULTAS_CONS$TMin, digits = 0)))
# names(IDADE_CAT_Tmin_freq) <- c("CAP","Idade","TMin","Freq")
# IDADE_CAT_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Ins))
# names(IDADE_CAT_Ins_freq) <- c("CAP","Idade","Ins","Freq")
# IDADE_CAT_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
# names(IDADE_CAT_EvapPiche_freq) <- c("CAP","Idade","EvapPiche","Freq")
# IDADE_CAT_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
# names(IDADE_CAT_TempCompMed_freq) <- c("CAP","Idade","TempCompMed","Freq")
# IDADE_CAT_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
# names(IDADE_CAT_UmidRelMed_freq) <- c("CAP","Idade","UmidRelMed","Freq")
# IDADE_CAT_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
# names(IDADE_CAT_VelVentMed_freq) <- c("CAP","Idade","VelVentMed","Freq")
# IDADE_CAT_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
# names(IDADE_CAT_TDelta_freq) <- c("CAP","Idade","TDelta","Freq")
```


```{r}
nrow(CONSULTAS_CONS)
```

```{r}
CAT_EvapPiche_freq$FreqRel <- CAT_EvapPiche_freq$Freq/nrow(CONSULTAS_CONS)
CAT_Ins_freq$FreqRel <- CAT_Ins_freq$Freq/nrow(CONSULTAS_CONS)
CAT_Precip_freq$FreqRel <- CAT_Precip_freq$Freq/nrow(CONSULTAS_CONS)
CAT_TDelta_freq$FreqRel <- CAT_TDelta_freq$Freq/nrow(CONSULTAS_CONS)
CAT_TDeltaA_freq$FreqRel <- CAT_TDeltaA_freq$Freq/nrow(CONSULTAS_CONS)
CAT_TempCompMed_freq$FreqRel <- CAT_TempCompMed_freq$Freq/nrow(CONSULTAS_CONS)
CAT_TMax_freq$FreqRel <- CAT_TMax_freq$Freq/nrow(CONSULTAS_CONS)
CAT_Tmin_freq$FreqRel <- CAT_Tmin_freq$Freq/nrow(CONSULTAS_CONS)
CAT_UmidRelMed_freq$FreqRel <- CAT_UmidRelMed_freq$Freq/nrow(CONSULTAS_CONS)
CAT_VelVentMed_freq$FreqRel <- CAT_VelVentMed_freq$Freq/nrow(CONSULTAS_CONS)

```

```{r}
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Precip)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMax)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMin)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Ins)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$EvapPiche)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TempCompMedia)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$UmidRelMed)
cor(CONSULTAS_CONS$CAP,CONSULTAS_CONS$VelVentMed)
cor(CONSULTAS_CONS$CAP,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Precip)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMax)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMin)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Ins)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$EvapPiche)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TempCompMedia)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$UmidRelMed)
var(CONSULTAS_CONS$CAP,CONSULTAS_CONS$VelVentMed)
var(CONSULTAS_CONS$CAP,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Precip)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMax)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMin)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$Ins)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$EvapPiche)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$TempCompMedia)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$UmidRelMed)
cov(CONSULTAS_CONS$CAP,CONSULTAS_CONS$VelVentMed)
cov(CONSULTAS_CONS$CAP,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$Precip)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$TMax)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$TMin)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$Ins)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$EvapPiche)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$TempCompMedia)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$UmidRelMed)
lm(CONSULTAS_CONS$CAP~CONSULTAS_CONS$VelVentMed)
lm(CONSULTAS_CONS$CAP~(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
lm(CAT_EvapPiche_freq$Freq~CAT_EvapPiche_freq$EvapPiche)
lm(CAT_Ins_freq$Freq~CAT_Ins_freq$Ins)
lm(CAT_Precip_freq$Freq~CAT_Precip_freq$Precip)
lm(CAT_TDelta_freq$Freq~CAT_TDelta_freq$TDelta)
lm(CAT_TDeltaA_freq$Freq~CAT_TDeltaA_freq$TDelta)
lm(CAT_TempCompMed_freq$Freq~CAT_TempCompMed_freq$TempCompMedia)
lm(CAT_TMax_freq$Freq~CAT_TMax_freq$TMax)
lm(CAT_Tmin_freq$Freq~CAT_Tmin_freq$TMin)
lm(CAT_UmidRelMed_freq$Freq~CAT_UmidRelMed_freq$UmidRelMed)
lm(CAT_VelVentMed_freq$Freq~CAT_VelVentMed_freq$VelVentMed)
```


```{r}
CAP_EvapPiche.lm <- lm(CAT_EvapPiche_freq$FreqRel~CAT_EvapPiche_freq$CAP + CAT_EvapPiche_freq$EvapPiche)
CAT_Ins_freq.lm <- lm(CAT_Ins_freq$Freq~CAT_Ins_freq$CAP + CAT_Ins_freq$Ins)
CAT_Precip_freq.lm <- lm(CAT_Precip_freq$Freq~CAT_Precip_freq$CAP + CAT_Precip_freq$Precip)
CAT_TDelta_freq.lm <- lm(CAT_TDelta_freq$Freq~CAT_TDelta_freq$CAP + CAT_TDelta_freq$TDelta)
CAT_TDeltaA_freq.lm <- lm(CAT_TDeltaA_freq$Freq~CAT_TDeltaA_freq$CAP + CAT_TDeltaA_freq$TDelta)
CAT_TempCompMed_freq.lm <- lm(CAT_TempCompMed_freq$Freq~CAT_TempCompMed_freq$CAP + CAT_TempCompMed_freq$TempCompMedia)
CAT_TMax_freq.lm <- lm(CAT_TMax_freq$Freq~CAT_TMax_freq$CAP + CAT_TMax_freq$TMax)
CAT_Tmin_freq.lm <- lm(CAT_Tmin_freq$Freq~CAT_Tmin_freq$CAP + CAT_Tmin_freq$TMin)
CAT_UmidRelMed_freq.lm <- lm(CAT_UmidRelMed_freq$Freq~CAT_UmidRelMed_freq$CAP + CAT_UmidRelMed_freq$UmidRelMed)
CAT_VelVentMed_freq.lm <- lm(CAT_VelVentMed_freq$Freq~CAT_VelVentMed_freq$CAP + CAT_VelVentMed_freq$VelVentMed)

```

```{r}
cooks.distance(CAP_EvapPiche.lm)
cooks.distance(CAT_Ins_freq.lm)
cooks.distance(CAT_Precip_freq.lm)
cooks.distance(CAT_TDelta_freq.lm)
cooks.distance(CAT_TDeltaA_freq.lm)
cooks.distance(CAT_TempCompMed_freq.lm)
cooks.distance(CAT_TMax_freq.lm)
cooks.distance(CAT_Tmin_freq.lm)
cooks.distance(CAT_UmidRelMed_freq.lm)
cooks.distance(CAT_VelVentMed_freq.lm)

```


```{r}
CAP_EvapPiche.im <- influence.measures(CAP_EvapPiche.lm)
CAT_Ins_freq.im <- influence.measures(CAT_Ins_freq.lm)
CAT_Precip_freq.im <- influence.measures(CAT_Precip_freq.lm)
CAT_TDelta_freq.im <- influence.measures(CAT_TDelta_freq.lm)
CAT_TDeltaA_freq.im <- influence.measures(CAT_TDeltaA_freq.lm)
CAT_TempCompMed_freq.im <- influence.measures(CAT_TempCompMed_freq.lm)
CAT_TMax_freq.im <- influence.measures(CAT_TMax_freq.lm)
CAT_Tmin_freq.im <- influence.measures(CAT_Tmin_freq.lm)
CAT_UmidRelMed_freq.im <- influence.measures(CAT_UmidRelMed_freq.lm)
CAT_VelVentMed_freq.im <- influence.measures(CAT_VelVentMed_freq.lm)

```

```{r}
summary(CAP_EvapPiche.im)
```
