==> TCC/4.AnaliseDosDados-SerieTemporal-CAP.Rmd <==
---
title: "Trabalho de Conclusão de Cusros - FGV T4 Big Data e Analytics"
author: "Ricardo Squassina Lee, Rafael Furlan, Djalma Gomes"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
  word_document: default
---

---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


---
```{r Inicia Ambiente, message=FALSE, include=FALSE}
#Iniciando o ambiente
rm(list = ls())
cat("\014") 
# setwd("~/Tresors/zz-pessoal/FGV/git/TCC")
setwd("./")
```

---
```{r Carrega bibliotecas necessarias, message=FALSE, warning=TRUE, include=FALSE}
if(!require(data.table)) {
  install.packages("data.table")
  library(data.table)
}

```

```{r}
CONSULTAS_CONS <- fread("consultas_cons_cap.csv", drop = c("V1","V12","V13"))
```


```{r MapReduce1}
Reduce_Data_CAP_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP))
names(Reduce_Data_CAP_freq) <- c("Data","CAP","Freq")

Reduce_Data_SEX_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$SEXO_BENEFICIARIO))
names(Reduce_Data_SEX_freq) <- c("Data","Sexo","Freq")

Reduce_Data_IDADE_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$IDADE_BENEFICIARIO))
names(Reduce_Data_IDADE_freq) <- c("Data","Idade","Freq")


```

```{r}

Reduce_Data_CAP_SEX_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$CAP))
names(Reduce_Data_CAP_SEX_freq) <- c("Data","Sexo","CAP","Freq")

Reduce_Data_CAP_IDADE_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$CAP))
names(Reduce_Data_CAP_IDADE_freq) <- c("Data","Idade","CAP","Freq")


```

```{r}

Reduce_Data_Precip_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$CAP,CONSULTAS_CONS$Precip))
names(Reduce_Data_Precip_freq) <- c("Data","CAP","Precip","Freq")

Reduce_Data_TMax_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMax))
names(Reduce_Data_TMax_freq) <- c("Data","CAP","TMax","Freq")

Reduce_Data_Tmin_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$TMin))
names(Reduce_Data_Tmin_freq) <- c("Data","CAP","TMin","Freq")

Reduce_Data_Ins_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$Ins))
names(Reduce_Data_Ins_freq) <- c("Data","CAP","Ins","Freq")

Reduce_Data_EvapPiche_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$EvapPiche))
names(Reduce_Data_EvapPiche_freq) <- c("Data","CAP","EvapPiche","Freq")

Reduce_Data_TempCompMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$TempCompMedia))
names(Reduce_Data_TempCompMed_freq) <- c("Data","CAP","TempCompMedia","Freq")

Reduce_Data_UmidRelMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$UmidRelMed))
names(Reduce_Data_UmidRelMed_freq) <- c("Data","CAP","UmidRelMed","Freq")

Reduce_Data_VelVentMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$VelVentMed))
names(Reduce_Data_VelVentMed_freq) <- c("Data","CAP","VelVentMed","Freq")

Reduce_Data_TempDelta_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
names(Reduce_Data_TempDelta_freq) <- c("Data","CAP","TDelta","Freq")


```

```{r}
Reduce_SEX_Data_Precip_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Precip))
names(Reduce_SEX_Data_Precip_freq) <- c("Data","CAP","Sexo","Precip","Freq")

Reduce_SEX_Data_TMax_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$TMax))
names(Reduce_SEX_Data_TMax_freq) <- c("Data","CAP","Sexo","TMax","Freq")

Reduce_SEX_Data_Tmin_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$TMin))
names(Reduce_SEX_Data_Tmin_freq) <- c("Data","CAP","Sexo","TMin","Freq")

Reduce_SEX_Data_Ins_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Ins))
names(Reduce_SEX_Data_Ins_freq) <- c("Data","CAP","Sexo","Ins","Freq")

Reduce_SEX_Data_EvapPiche_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
names(Reduce_SEX_Data_EvapPiche_freq) <- c("Data","CAP","Sexo","EvapPiche","Freq")

Reduce_SEX_Data_TempCompMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
names(Reduce_SEX_Data_TempCompMed_freq) <- c("Data","CAP","Sexo","TempCompMedia","Freq")

Reduce_SEX_Data_UmidRelMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
names(Reduce_SEX_Data_UmidRelMed_freq) <- c("Data","CAP","Sexo","UmidRelMed","Freq")

Reduce_SEX_Data_VelVentMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
names(Reduce_SEX_Data_VelVentMed_freq) <- c("Data","CAP","Sexo","VelVentMed","Freq")

Reduce_SEX_Data_TempDelta_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$SEXO_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
names(Reduce_SEX_Data_TempDelta_freq) <- c("Data","CAP","Sexo","TDelta","Freq")


```

```{r}


Reduce_Data_IDADE_Precip_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Precip))
names(Reduce_Data_IDADE_Precip_freq) <- c("Data","CAP","Idade","Precip","Freq")

Reduce_Data_IDADE_TMax_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$TMax))
names(Reduce_Data_IDADE_TMax_freq) <- c("Data","CAP","idade","TMax","Freq")

Reduce_Data_IDADE_Tmin_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$TMin))
names(Reduce_Data_IDADE_Tmin_freq) <- c("Data","CAP","Idade","TMin","Freq")

Reduce_Data_IDADE_Ins_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Ins))
names(Reduce_Data_IDADE_Ins_freq) <- c("Data","CAP","Idade","Ins","Freq")

Reduce_Data_IDADE_EvapPiche_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
names(Reduce_Data_IDADE_EvapPiche_freq) <- c("Data","CAP","Idade","EvapPiche","Freq")

Reduce_Data_IDADE_TempCompMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
names(Reduce_Data_IDADE_TempCompMed_freq) <- c("Data","CAP","Idade","TempCompMedia","Freq")

Reduce_Data_IDADE_UmidRelMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
names(Reduce_Data_IDADE_UmidRelMed_freq) <- c("Data","CAP","Idade","UmidRelMed","Freq")

Reduce_Data_IDADE_VelVentMed_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
names(Reduce_Data_IDADE_VelVentMed_freq) <- c("Data","CAP","Idade","VelVentMed","Freq")

Reduce_Data_IDADE_TempDelta_freq <- as.data.frame(table(substr(CONSULTAS_CONS$DT_INTERNACAO,1,7),CONSULTAS_CONS$CAP,CONSULTAS_CONS$IDADE_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
names(Reduce_Data_IDADE_TempDelta_freq) <- c("Data","CAP","Idade","TDelta","Freq")
```


```{r}
CONSULTAS_CONS_LN <- NULL
CONSULTAS_CONS_LN <- log(CONSULTAS_CONS[5:12]/base::rbind(NA,CONSULTAS_CONS[5:12][-nrow(CONSULTAS_CONS[5:12]),]))
CONSULTAS_CONS_LN <- CONSULTAS_CONS_LN[-1,]
```
