==> TCC/13.AnaliseDosDadosDataDefasada-MR-COR-VAR-COV-CAP.Rmd <==
---
title: "Trabalho de Conclusão de Cusros - FGV T4 Big Data e Analytics"
author: "Ricardo Squassina Lee, Rafael Furlan, Djalma Gomes"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
  word_document: default
---

---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


---
```{r Inicia Ambiente, message=FALSE, include=FALSE}
#Iniciando o ambiente
rm(list = ls())
cat("\014") 
# setwd("~/Tresors/zz-pessoal/FGV/git/TCC")
setwd("./")
```

---
```{r Carrega bibliotecas necessarias, message=FALSE, warning=TRUE, include=FALSE}
if(!require(data.table)) {
  install.packages("data.table")
  library(data.table)
}

```

```{r}
CONSULTAS_CONS <- fread("consultas_cons7.csv", drop = c("V1"))
```

# MapReduce

```{r MapReduce}
CID_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$Precip))
names(CID_Precip_freq) <- c("CID","Precip","Freq")
CID_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CID,round(CONSULTAS_CONS$TMax, digits = 0)))
names(CID_TMax_freq) <- c("CID","TMax","Freq")
CID_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CID,round(CONSULTAS_CONS$TMin, digits = 0)))
names(CID_Tmin_freq) <- c("CID","TMin","Freq")
CID_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$Ins))
names(CID_Ins_freq) <- c("CID","Ins","Freq")
CID_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$EvapPiche))
names(CID_EvapPiche_freq) <- c("CID","EvapPiche","Freq")
CID_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$TempCompMedia))
names(CID_TempCompMed_freq) <- c("CID","TempCompMedia","Freq")
CID_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$UmidRelMed))
names(CID_UmidRelMed_freq) <- c("CID","UmidRelMed","Freq")
CID_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$VelVentMed))
names(CID_VelVentMed_freq) <- c("CID","VelVentMed","Freq")
CID_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CID,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
names(CID_TDelta_freq) <- c("CID","TDelta","Freq")
CID_TDeltaA_freq <- as.data.frame(table(CONSULTAS_CONS$CID,round((CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin), digits = 0)))
names(CID_TDeltaA_freq) <- c("CID","TDelta","Freq")

# SEX_CID_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Precip))
# names(SEX_CID_Precip_freq) <- c("CID","Sex","Freq")
# SEX_CID_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,round(CONSULTAS_CONS$TMax, digits = 0)))
# names(SEX_CID_TMax_freq) <- c("CID","Sex","Tmax","Freq")
# SEX_CID_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,round(CONSULTAS_CONS$TMin, digits =0 )))
# names(SEX_CID_Tmin_freq) <- c("CID","Sex","TMin","Freq")
# SEX_CID_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$Ins))
# names(SEX_CID_Ins_freq) <- c("CID","Sex","Ins","Freq")
# SEX_CID_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
# names(SEX_CID_EvapPiche_freq) <- c("CID","Sex","EvapPiche","Freq")
# SEX_CID_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
# names(SEX_CID_TempCompMed_freq) <- c("CID","Sex","TempCompMed","Freq")
# SEX_CID_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
# names(SEX_CID_UmidRelMed_freq) <- c("CID","Sex","UmidRelMed","Freq")
# SEX_CID_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
# names(SEX_CID_VelVentMed_freq) <- c("CID","Sex","VelVentMed","Freq")
# SEX_CID_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$SEXO_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
# names(SEX_CID_TDelta_freq) <- c("CID","Sex","TDelta","Freq")
# 
# IDADE_CID_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Precip))
# names(IDADE_CID_Precip_freq) <- c("CID","Idade","Precip","Freq")
# IDADE_CID_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,round(CONSULTAS_CONS$TMax, digits = 0)))
# names(IDADE_CID_TMax_freq) <- c("CID","Idade","TMax","Freq")
# IDADE_CID_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,round(CONSULTAS_CONS$TMin, digits = 0)))
# names(IDADE_CID_Tmin_freq) <- c("CID","Idade","TMin","Freq")
# IDADE_CID_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$Ins))
# names(IDADE_CID_Ins_freq) <- c("CID","Idade","Ins","Freq")
# IDADE_CID_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$EvapPiche))
# names(IDADE_CID_EvapPiche_freq) <- c("CID","Idade","EvapPiche","Freq")
# IDADE_CID_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$TempCompMedia))
# names(IDADE_CID_TempCompMed_freq) <- c("CID","Idade","TempCompMed","Freq")
# IDADE_CID_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$UmidRelMed))
# names(IDADE_CID_UmidRelMed_freq) <- c("CID","Idade","UmidRelMed","Freq")
# IDADE_CID_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,CONSULTAS_CONS$VelVentMed))
# names(IDADE_CID_VelVentMed_freq) <- c("CID","Idade","VelVentMed","Freq")
# IDADE_CID_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CID,CONSULTAS_CONS$IDADE_BENEFICIARIO,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin)))
# names(IDADE_CID_TDelta_freq) <- c("CID","Idade","TDelta","Freq")
```


```{r}
nrow(CONSULTAS_CONS)
```

```{r}
CID_EvapPiche_freq$FreqRel <- CID_EvapPiche_freq$Freq/nrow(CONSULTAS_CONS)
CID_Ins_freq$FreqRel <- CID_Ins_freq$Freq/nrow(CONSULTAS_CONS)
CID_Precip_freq$FreqRel <- CID_Precip_freq$Freq/nrow(CONSULTAS_CONS)
CID_TDelta_freq$FreqRel <- CID_TDelta_freq$Freq/nrow(CONSULTAS_CONS)
CID_TDeltaA_freq$FreqRel <- CID_TDeltaA_freq$Freq/nrow(CONSULTAS_CONS)
CID_TempCompMed_freq$FreqRel <- CID_TempCompMed_freq$Freq/nrow(CONSULTAS_CONS)
CID_TMax_freq$FreqRel <- CID_TMax_freq$Freq/nrow(CONSULTAS_CONS)
CID_Tmin_freq$FreqRel <- CID_Tmin_freq$Freq/nrow(CONSULTAS_CONS)
CID_UmidRelMed_freq$FreqRel <- CID_UmidRelMed_freq$Freq/nrow(CONSULTAS_CONS)
CID_VelVentMed_freq$FreqRel <- CID_VelVentMed_freq$Freq/nrow(CONSULTAS_CONS)

```

```{r}
cor(as.numeric(CID_EvapPiche_freq$EvapPiche),CID_EvapPiche_freq$Freq)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMax)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMin)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$Ins)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$EvapPiche)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$TempCompMedia)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$UmidRelMed)
cor(CONSULTAS_CONS$CID,CONSULTAS_CONS$VelVentMed)
cor(CONSULTAS_CONS$CID,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$Precip)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMax)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMin)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$Ins)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$EvapPiche)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$TempCompMedia)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$UmidRelMed)
var(CONSULTAS_CONS$CID,CONSULTAS_CONS$VelVentMed)
var(CONSULTAS_CONS$CID,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$Precip)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMax)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$TMin)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$Ins)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$EvapPiche)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$TempCompMedia)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$UmidRelMed)
cov(CONSULTAS_CONS$CID,CONSULTAS_CONS$VelVentMed)
cov(CONSULTAS_CONS$CID,(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$Precip)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$TMax)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$TMin)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$Ins)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$EvapPiche)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$TempCompMedia)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$UmidRelMed)
lm(CONSULTAS_CONS$CID~CONSULTAS_CONS$VelVentMed)
lm(CONSULTAS_CONS$CID~(CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin))

```

```{r}
lm(CID_EvapPiche_freq$Freq~CID_EvapPiche_freq$EvapPiche)
lm(CID_Ins_freq$Freq~CID_Ins_freq$Ins)
lm(CID_Precip_freq$Freq~CID_Precip_freq$Precip)
lm(CID_TDelta_freq$Freq~CID_TDelta_freq$TDelta)
lm(CID_TDeltaA_freq$Freq~CID_TDeltaA_freq$TDelta)
lm(CID_TempCompMed_freq$Freq~CID_TempCompMed_freq$TempCompMedia)
lm(CID_TMax_freq$Freq~CID_TMax_freq$TMax)
lm(CID_Tmin_freq$Freq~CID_Tmin_freq$TMin)
lm(CID_UmidRelMed_freq$Freq~CID_UmidRelMed_freq$UmidRelMed)
lm(CID_VelVentMed_freq$Freq~CID_VelVentMed_freq$VelVentMed)
```


```{r}
CID_EvapPiche.lm <- lm(CID_EvapPiche_freq$FreqRel~CID_EvapPiche_freq$CID + CID_EvapPiche_freq$EvapPiche)
CID_Ins_freq.lm <- lm(CID_Ins_freq$Freq~CID_Ins_freq$CID + CID_Ins_freq$Ins)
CID_Precip_freq.lm <- lm(CID_Precip_freq$Freq~CID_Precip_freq$CID + CID_Precip_freq$Precip)
CID_TDelta_freq.lm <- lm(CID_TDelta_freq$Freq~CID_TDelta_freq$CID + CID_TDelta_freq$TDelta)
CID_TDeltaA_freq.lm <- lm(CID_TDeltaA_freq$Freq~CID_TDeltaA_freq$CID + CID_TDeltaA_freq$TDelta)
CID_TempCompMed_freq.lm <- lm(CID_TempCompMed_freq$Freq~CID_TempCompMed_freq$CID + CID_TempCompMed_freq$TempCompMedia)
CID_TMax_freq.lm <- lm(CID_TMax_freq$Freq~CID_TMax_freq$CID + CID_TMax_freq$TMax)
CID_Tmin_freq.lm <- lm(CID_Tmin_freq$Freq~CID_Tmin_freq$CID + CID_Tmin_freq$TMin)
CID_UmidRelMed_freq.lm <- lm(CID_UmidRelMed_freq$Freq~CID_UmidRelMed_freq$CID + CID_UmidRelMed_freq$UmidRelMed)
CID_VelVentMed_freq.lm <- lm(CID_VelVentMed_freq$Freq~CID_VelVentMed_freq$CID + CID_VelVentMed_freq$VelVentMed)

```

```{r}
cooks.distance(CID_EvapPiche.lm)
cooks.distance(CID_Ins_freq.lm)
cooks.distance(CID_Precip_freq.lm)
cooks.distance(CID_TDelta_freq.lm)
cooks.distance(CID_TDeltaA_freq.lm)
cooks.distance(CID_TempCompMed_freq.lm)
cooks.distance(CID_TMax_freq.lm)
cooks.distance(CID_Tmin_freq.lm)
cooks.distance(CID_UmidRelMed_freq.lm)
cooks.distance(CID_VelVentMed_freq.lm)

```


```{r}
CID_EvapPiche.im <- influence.measures(CID_EvapPiche.lm)
CID_Ins_freq.im <- influence.measures(CID_Ins_freq.lm)
CID_Precip_freq.im <- influence.measures(CID_Precip_freq.lm)
CID_TDelta_freq.im <- influence.measures(CID_TDelta_freq.lm)
CID_TDeltaA_freq.im <- influence.measures(CID_TDeltaA_freq.lm)
CID_TempCompMed_freq.im <- influence.measures(CID_TempCompMed_freq.lm)
CID_TMax_freq.im <- influence.measures(CID_TMax_freq.lm)
CID_Tmin_freq.im <- influence.measures(CID_Tmin_freq.lm)
CID_UmidRelMed_freq.im <- influence.measures(CID_UmidRelMed_freq.lm)
CID_VelVentMed_freq.im <- influence.measures(CID_VelVentMed_freq.lm)

```

```{r}
summary(CID_EvapPiche.im)
```
