==> TCC/14.AnaliseDosDados-Graph-CAP.Rmd <==
---
title: "Trabalho de Conclusão de Cusros - FGV T4 Big Data e Analytics"
author: "Ricardo Squassina Lee, Rafael Furlan, Djalma Gomes"
output:
  word_document: default
  html_notebook: default
  pdf_document: default
  html_document:
    df_print: paged
---

---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


---
```{r Inicia Ambiente, message=FALSE, include=FALSE}
#Iniciando o ambiente
rm(list = ls())
cat("\014") 
# setwd("~/Tresors/zz-pessoal/FGV/git/TCC")
setwd("./")
```

---
```{r Carrega bibliotecas necessarias, message=FALSE, warning=TRUE, include=FALSE}
if(!require(data.table)) {
  install.packages("data.table")
  library(data.table)
}

if(!require(network)) {
  install.packages("network")
  library(network)
}

if(!require(sna)) {
  install.packages("sna")
  library(sna)
}

if(!require(rgl)) {
  install.packages("rgl")
  library(rgl)
}

```

```{r}
CONSULTAS_CONS <- fread("consultas_cons_cap7.csv", drop = c("V1"))
```

# MapReduce

```{r MapReduce}
CAP_Precip_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$Precip, digits = 0)))
names(CAP_Precip_freq) <- c("CAP","Precip","Freq")

CAP_TMax_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$TMax, digits = 0)))
names(CAP_TMax_freq) <- c("CAP","TMax","Freq")

CAP_Tmin_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$TMin, digits = 0)))
names(CAP_Tmin_freq) <- c("CAP","TMin","Freq")

CAP_Ins_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$Ins, digits = 0)))
names(CAP_Ins_freq) <- c("CAP","Ins","Freq")

CAP_EvapPiche_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$EvapPiche, digits = 0)))
names(CAP_EvapPiche_freq) <- c("CAP","EvapPiche","Freq")

CAP_TempCompMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$TempCompMedia, digits = 0)))
names(CAP_TempCompMed_freq) <- c("CAP","TempCompMedia","Freq")

CAP_UmidRelMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$UmidRelMed, digits = 0)))
names(CAP_UmidRelMed_freq) <- c("CAP","UmidRelMed","Freq")

CAP_VelVentMed_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round(CONSULTAS_CONS$VelVentMed, digits = 0)))
names(CAP_VelVentMed_freq) <- c("CAP","VelVentMed","Freq")

CAP_TDelta_freq <- as.data.frame(table(CONSULTAS_CONS$CAP,round((CONSULTAS_CONS$TMax - CONSULTAS_CONS$TMin), digits = 0)))
names(CAP_TDelta_freq) <- c("CAP","TDelta","Freq")


```


```{r}
CAP_EvapPiche_freq$FreqRel <- CAP_EvapPiche_freq$Freq/nrow(CONSULTAS_CONS)
CAP_Ins_freq$FreqRel <- CAP_Ins_freq$Freq/nrow(CONSULTAS_CONS)
CAP_Precip_freq$FreqRel <- CAP_Precip_freq$Freq/nrow(CONSULTAS_CONS)
CAP_TDelta_freq$FreqRel <- CAP_TDelta_freq$Freq/nrow(CONSULTAS_CONS)
CAP_TempCompMed_freq$FreqRel <- CAP_TempCompMed_freq$Freq/nrow(CONSULTAS_CONS)
CAP_TMax_freq$FreqRel <- CAP_TMax_freq$Freq/nrow(CONSULTAS_CONS)
CAP_Tmin_freq$FreqRel <- CAP_Tmin_freq$Freq/nrow(CONSULTAS_CONS)
CAP_UmidRelMed_freq$FreqRel <- CAP_UmidRelMed_freq$Freq/nrow(CONSULTAS_CONS)
CAP_VelVentMed_freq$FreqRel <- CAP_VelVentMed_freq$Freq/nrow(CONSULTAS_CONS)

```



```{r}

nrow <- max(unique(as.integer(CAP_EvapPiche_freq$CAP)))
ncol <- max(unique(as.integer(CAP_EvapPiche_freq$EvapPiche)))

MATRIX_EVAPPICHE <- matrix(CAP_EvapPiche_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_EVAPPICHE) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_EVAPPICHE) <- unique(CAP_EvapPiche_freq$EvapPiche)

```

```{r}

nrow <- max(unique(as.integer(CAP_Ins_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Ins_freq$Ins)))

MATRIX_INS <- matrix(CAP_Ins_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_INS) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_INS) <- unique(CAP_Ins_freq$Ins)

```

```{r}

nrow <- max(unique(as.integer(CAP_Precip_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Precip_freq$Precip)))

MATRIX_PRECIP <- matrix(CAP_Precip_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_PRECIP) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_PRECIP) <- unique(CAP_Precip_freq$Precip)

```

```{r}

nrow <- max(unique(as.integer(CAP_TDelta_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TDelta_freq$TDelta)))

MATRIX_TDELTA <- matrix(CAP_TDelta_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_TDELTA) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TDELTA) <- unique(CAP_TDelta_freq$TDelta)

```

```{r}

nrow <- max(unique(as.integer(CAP_TempCompMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TempCompMed_freq$TempCompMedia)))

MATRIX_TEMPCOMPMED <- matrix(CAP_TempCompMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_TEMPCOMPMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TEMPCOMPMED) <- unique(CAP_TempCompMed_freq$TempCompMedia)

```

```{r}

nrow <- max(unique(as.integer(CAP_TMax_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TMax_freq$TMax)))

MATRIX_TMAX <- matrix(CAP_TMax_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_TMAX) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TMAX) <- unique(CAP_TMax_freq$TMax)


```

```{r}

nrow <- max(unique(as.integer(CAP_Tmin_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Tmin_freq$TMin)))

MATRIX_TMIN <- matrix(CAP_Tmin_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_TMIN) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TMIN) <- unique(CAP_Tmin_freq$TMin)


```

```{r}

nrow <- max(unique(as.integer(CAP_UmidRelMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_UmidRelMed_freq$UmidRelMed)))

MATRIX_UMIDRELMED <- matrix(CAP_UmidRelMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_UMIDRELMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_UMIDRELMED) <- unique(CAP_UmidRelMed_freq$UmidRelMed)

```

```{r}

nrow <- max(unique(as.integer(CAP_VelVentMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_VelVentMed_freq$VelVentMed)))

MATRIX_VELVENTMED <- matrix(CAP_VelVentMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_VELVENTMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_VELVENTMED) <- unique(CAP_VelVentMed_freq$VelVentMed)

```



```{r}

nrow <- max(unique(as.integer(CAP_EvapPiche_freq$CAP)))
ncol <- max(unique(as.integer(CAP_EvapPiche_freq$EvapPiche)))

MATRIX_EVAPPICHE_REL <- matrix(CAP_EvapPiche_freq$FreqRel, nrow = nrow, ncol = ncol)

rownames(MATRIX_EVAPPICHE_REL) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_EVAPPICHE_REL) <- unique(CAP_EvapPiche_freq$EvapPiche)

```

```{r}

nrow <- max(unique(as.integer(CAP_Ins_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Ins_freq$Ins)))

MATRIX_INS <- matrix(CAP_Ins_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_INS) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_INS) <- unique(CAP_Ins_freq$Ins)

```

```{r}

nrow <- max(unique(as.integer(CAP_Precip_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Precip_freq$Precip)))

MATRIX_PRECIP <- matrix(CAP_Precip_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_PRECIP) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_PRECIP) <- unique(CAP_Precip_freq$Precip)

```

```{r}

nrow <- max(unique(as.integer(CAP_TDelta_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TDelta_freq$TDelta)))

MATRIX_TDELTA <- matrix(CAP_TDelta_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_TDELTA) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TDELTA) <- unique(CAP_TDelta_freq$TDelta)

```

```{r}

nrow <- max(unique(as.integer(CAP_TempCompMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TempCompMed_freq$TempCompMedia)))

MATRIX_TEMPCOMPMED <- matrix(CAP_TempCompMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_TEMPCOMPMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TEMPCOMPMED) <- unique(CAP_TempCompMed_freq$TempCompMedia)

```

```{r}

nrow <- max(unique(as.integer(CAP_TMax_freq$CAP)))
ncol <- max(unique(as.integer(CAP_TMax_freq$TMax)))

MATRIX_TMAX <- matrix(CAP_TMax_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_TMAX) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TMAX) <- unique(CAP_TMax_freq$TMax)


```

```{r}

nrow <- max(unique(as.integer(CAP_Tmin_freq$CAP)))
ncol <- max(unique(as.integer(CAP_Tmin_freq$TMin)))

MATRIX_TMIN <- matrix(CAP_Tmin_freq$Freq, nrow = nrow, ncol = ncol)

rownames(MATRIX_TMIN) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_TMIN) <- unique(CAP_Tmin_freq$TMin)


```

```{r}

nrow <- max(unique(as.integer(CAP_UmidRelMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_UmidRelMed_freq$UmidRelMed)))

MATRIX_UMIDRELMED <- matrix(CAP_UmidRelMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_UMIDRELMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_UMIDRELMED) <- unique(CAP_UmidRelMed_freq$UmidRelMed)

```

```{r}

nrow <- max(unique(as.integer(CAP_VelVentMed_freq$CAP)))
ncol <- max(unique(as.integer(CAP_VelVentMed_freq$VelVentMed)))

MATRIX_VELVENTMED <- matrix(CAP_VelVentMed_freq$Freq, nrow = nrow, ncol = ncol)


rownames(MATRIX_VELVENTMED) <- c("CAP 1"
                                ,"CAP 2"
                                ,"CAP 3"
                                ,"CAP 4"
                                ,"CAP 5"
                                ,"CAP 6"
                                ,"CAP 7"
                                ,"CAP 8"
                                ,"CAP 9"
                                ,"CAP 10"
                                ,"CAP 11"
                                ,"CAP 12"
                                ,"CAP 13"
                                ,"CAP 14"
                                ,"CAP 15"
                                ,"CAP 16"
                                ,"CAP 17"
                                ,"CAP 18"
                                ,"CAP 19"
                                ,"CAP 20"
                                ,"CAP 21"
                                ,"CAP 22")

colnames(MATRIX_VELVENTMED) <- unique(CAP_VelVentMed_freq$VelVentMed)

```




Análises Grafos - Evaporação do Piche - Frequência Nominal

```{r}
dim(MATRIX_EVAPPICHE)
```

```{r}
sna::degree(MATRIX_EVAPPICHE,gmode="digraph",cmode="freeman")
```

```{r}
sna::closeness(MATRIX_EVAPPICHE,gmode="digraph")
```

```{r}
sna::betweenness(MATRIX_EVAPPICHE,gmode="digraph")
```

```{r}
sna::bicomponent.dist(MATRIX_EVAPPICHE)
```

```{r}
sna::bicomponent.dist(MATRIX_EVAPPICHE, symmetrize = c("strong", "weak"))
```

```{r}
sna::components(MATRIX_EVAPPICHE,connected="weak")
```

```{r}
sna::components(MATRIX_EVAPPICHE,connected="strong")
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE,gtrans,cmode="size") 
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE,gtrans,cmode="edges") 
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE,gtrans,cmode="dyad.census")
```

```{r}
sna::diag.remove(MATRIX_EVAPPICHE)
```

```{r}
sna::efficiency(MATRIX_EVAPPICHE)
```

```{r}
sna::gden(MATRIX_EVAPPICHE)
```

```{r}

gplot(MATRIX_EVAPPICHE,gmode="twomode",displaylabels = TRUE,
      edge.col="gray",label.cex = 0.7,usearrows=FALSE,
      vertex.cex = closeness(MATRIX_EVAPPICHE,gmode="twomode")*3)
```

```{r}

eq <- equiv.clust(MATRIX_EVAPPICHE)
plot(eq)
# gplot(diag.remove(MATRIX_EVAPPICHE))
```

```{r}

######################### SNA

# Find the Central Graph of a Labeled Graph Stack
cg <- centralgraph(MATRIX_EVAPPICHE)
gplot3d(cg)
```

```{r}

# Compute The Closeness Centrality Scores Of Network Positions
closeness(MATRIX_EVAPPICHE)
```

```{r}

# Find weak components
components(MATRIX_EVAPPICHE,connected="weak")
```

```{r}

# Find strong components
components(MATRIX_EVAPPICHE,connected="strong")
```

```{r}

# Compute Graph Efficiency Scores
efficiency(MATRIX_EVAPPICHE, g=NULL, diag=FALSE)
```

```{r}

# Compute The Degree Centrality Scores Of Network Positions
degree(MATRIX_EVAPPICHE, g=1, nodes=NULL, gmode="digraph", diag=FALSE, tmaxdev=FALSE, cmode="freeman", rescale=FALSE, ignore.eval=FALSE)
```

```{r}

# Find The Density Of A Graph
gden(MATRIX_EVAPPICHE)

```

Análises Grafos - Precipitação - Frequência Nominal

```{r}
dim(MATRIX_PRECIP)
```

```{r}
sna::degree(MATRIX_PRECIP,gmode="digraph",cmode="freeman")
```

```{r}
sna::closeness(MATRIX_PRECIP,gmode="digraph")
```

```{r}
sna::betweenness(MATRIX_PRECIP,gmode="digraph")
```

```{r}
sna::bicomponent.dist(MATRIX_PRECIP)
```

```{r}
sna::bicomponent.dist(MATRIX_PRECIP, symmetrize = c("strong", "weak"))
```

```{r}
sna::components(MATRIX_PRECIP,connected="weak")
```

```{r}
sna::components(MATRIX_PRECIP,connected="strong")
```

```{r}
sna::cug.test(MATRIX_PRECIP,gtrans,cmode="size") 
```

```{r}
sna::cug.test(MATRIX_PRECIP,gtrans,cmode="edges") 
```

```{r}
sna::cug.test(MATRIX_PRECIP,gtrans,cmode="dyad.census")
```

```{r}
sna::diag.remove(MATRIX_PRECIP)
```

```{r}
sna::efficiency(MATRIX_PRECIP)
```

```{r}
sna::gden(MATRIX_PRECIP)
```

```{r}

gplot(MATRIX_PRECIP,gmode="twomode",displaylabels = TRUE,
      edge.col="gray",label.cex = 0.7,usearrows=FALSE,
      vertex.cex = closeness(MATRIX_PRECIP,gmode="twomode")*3)
```

```{r}

eq <- equiv.clust(MATRIX_PRECIP)
plot(eq)
# gplot(diag.remove(MATRIX_PRECIP))
```

```{r}

######################### SNA

# Find the Central Graph of a Labeled Graph Stack
cg <- centralgraph(MATRIX_PRECIP)
gplot3d(cg)
```

```{r}

# Compute The Closeness Centrality Scores Of Network Positions
closeness(MATRIX_PRECIP)
```

```{r}

# Find weak components
components(MATRIX_PRECIP,connected="weak")
```

```{r}

# Find strong components
components(MATRIX_PRECIP,connected="strong")
```

```{r}

# Compute Graph Efficiency Scores
efficiency(MATRIX_PRECIP, g=NULL, diag=FALSE)
```

```{r}

# Compute The Degree Centrality Scores Of Network Positions
degree(MATRIX_PRECIP, g=1, nodes=NULL, gmode="digraph", diag=FALSE, tmaxdev=FALSE, cmode="freeman", rescale=FALSE, ignore.eval=FALSE)
```

```{r}

# Find The Density Of A Graph
gden(MATRIX_PRECIP)

```


Análises Grafos - Variação de Tempertura (Delta) - Frequência Nominal

```{r}
dim(MATRIX_TDELTA)
```

```{r}
sna::degree(MATRIX_TDELTA,gmode="digraph",cmode="freeman")
```

```{r}
sna::closeness(MATRIX_TDELTA,gmode="digraph")
```

```{r}
sna::betweenness(MATRIX_TDELTA,gmode="digraph")
```

```{r}
sna::bicomponent.dist(MATRIX_TDELTA)
```

```{r}
sna::bicomponent.dist(MATRIX_TDELTA, symmetrize = c("strong", "weak"))
```

```{r}
sna::components(MATRIX_TDELTA,connected="weak")
```

```{r}
sna::components(MATRIX_TDELTA,connected="strong")
```

```{r}
sna::cug.test(MATRIX_TDELTA,gtrans,cmode="size") 
```

```{r}
sna::cug.test(MATRIX_TDELTA,gtrans,cmode="edges") 
```

```{r}
sna::cug.test(MATRIX_TDELTA,gtrans,cmode="dyad.census")
```

```{r}
sna::diag.remove(MATRIX_TDELTA)
```

```{r}
sna::efficiency(MATRIX_TDELTA)
```

```{r}
sna::gden(MATRIX_TDELTA)
```

```{r}

gplot(MATRIX_TDELTA,gmode="twomode",displaylabels = TRUE,
      edge.col="gray",label.cex = 0.7,usearrows=FALSE,
      vertex.cex = closeness(MATRIX_TDELTA,gmode="twomode")*3)
```

```{r}

eq <- equiv.clust(MATRIX_TDELTA)
plot(eq)
# gplot(diag.remove(MATRIX_TDELTA))
```

```{r}

######################### SNA

# Find the Central Graph of a Labeled Graph Stack
cg <- centralgraph(MATRIX_TDELTA)
gplot3d(cg)
```

```{r}

# Compute The Closeness Centrality Scores Of Network Positions
closeness(MATRIX_TDELTA)
```

```{r}

# Find weak components
components(MATRIX_TDELTA,connected="weak")
```

```{r}

# Find strong components
components(MATRIX_TDELTA,connected="strong")
```

```{r}

# Compute Graph Efficiency Scores
efficiency(MATRIX_TDELTA, g=NULL, diag=FALSE)
```

```{r}

# Compute The Degree Centrality Scores Of Network Positions
degree(MATRIX_TDELTA, g=1, nodes=NULL, gmode="digraph", diag=FALSE, tmaxdev=FALSE, cmode="freeman", rescale=FALSE, ignore.eval=FALSE)
```

```{r}

# Find The Density Of A Graph
gden(MATRIX_TDELTA)

```


Análises Grafos - Evaporação do Piche - Frequência Relativa

```{r}
dim(MATRIX_EVAPPICHE_REL)
```

```{r}
sna::degree(MATRIX_EVAPPICHE_REL,gmode="digraph",cmode="freeman")
```

```{r}
sna::closeness(MATRIX_EVAPPICHE_REL,gmode="digraph")
```

```{r}
sna::betweenness(MATRIX_EVAPPICHE_REL,gmode="digraph")
```

```{r}
sna::bicomponent.dist(MATRIX_EVAPPICHE_REL)
```

```{r}
sna::bicomponent.dist(MATRIX_EVAPPICHE_REL, symmetrize = c("strong", "weak"))
```

```{r}
sna::components(MATRIX_EVAPPICHE_REL,connected="weak")
```

```{r}
sna::components(MATRIX_EVAPPICHE_REL,connected="strong")
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE_REL,gtrans,cmode="size") 
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE_REL,gtrans,cmode="edges") 
```

```{r}
sna::cug.test(MATRIX_EVAPPICHE_REL,gtrans,cmode="dyad.census")
```

```{r}
sna::diag.remove(MATRIX_EVAPPICHE_REL)
```

```{r}
sna::efficiency(MATRIX_EVAPPICHE_REL)
```

```{r}
sna::gden(MATRIX_EVAPPICHE_REL)
```

```{r}

gplot(MATRIX_EVAPPICHE_REL,gmode="twomode",displaylabels = TRUE,
      edge.col="gray",label.cex = 0.7,usearrows=FALSE,
      vertex.cex = closeness(MATRIX_EVAPPICHE_REL,gmode="twomode")*3)

```

```{r}

eq <- equiv.clust(MATRIX_EVAPPICHE_REL)
plot(eq)
# gplot(diag.remove(MATRIX_EVAPPICHE_REL))

```

```{r}
######################### SNA

# Find the Central Graph of a Labeled Graph Stack
cg <- centralgraph(MATRIX_EVAPPICHE_REL)
gplot3d(cg)

```

```{r}
# Compute The Closeness Centrality Scores Of Network Positions
closeness(MATRIX_EVAPPICHE_REL)

```

```{r}
# Find weak components
components(MATRIX_EVAPPICHE_REL,connected="weak")

```

```{r}
# Find strong components
components(MATRIX_EVAPPICHE_REL,connected="strong")

```

```{r}
# Compute Graph Efficiency Scores
efficiency(MATRIX_EVAPPICHE_REL, g=NULL, diag=FALSE)

```

```{r}
# Compute The Degree Centrality Scores Of Network Positions
degree(MATRIX_EVAPPICHE_REL, g=1, nodes=NULL, gmode="digraph", diag=FALSE, tmaxdev=FALSE, cmode="freeman", rescale=FALSE, ignore.eval=FALSE)

```

```{r}
# Find The Density Of A Graph
gden(MATRIX_EVAPPICHE_REL)

```
